requirements.txt
bcrypt>=3.2.0
pytz>=2021.1
kafka-python>=2.0.1
